# src/domain/follows/dto/follower-users-response.dto.ts
export class FollowerPreviewDto {
  id!: number;
  username!: string;
  profileImageUrl!: string | null;
  isFollowing!: boolean;
}

export class FollowersResponseDto {
  items!: FollowerPreviewDto[];
  nextCursor!: string | null;
  hasNext!: boolean;
}