# src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"FollowerPreviewDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"profileImageUrl": {
						"type": "string",
						"nullable": true
					},
					"isFollowing": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"username",
					"profileImageUrl",
					"isFollowing"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FollowersResponseDto": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/FollowerPreviewDto"
						},
						"type": "array"
					},
					"nextCursor": {
						"type": "string",
						"nullable": true
					},
					"hasNext": {
						"type": "boolean"
					}
				},
				"required": [
					"items",
					"nextCursor",
					"hasNext"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SignUpDto": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"passwordConfirm": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password",
					"passwordConfirm"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CheckIdResponse": {
				"properties": {
					"isAvaliable": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"isAvaliable",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignInResponseDto": {
				"properties": {
					"accessToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignInDto": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ovly-express",
		"version": "1.0.0",
		"description": "express mjs ts",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "ripetangerine",
			"url": "kwak jk"
		}
	},
	"paths": {
...
		"/follows/{userId}": {
			"post": {
				"operationId": "ToggleFollow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isFollowing": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"isFollowing",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Follows"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/follows/followers/{viewerId}": {
			"get": {
				"operationId": "GetFollowers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FollowersResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "viewerId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/follows/followings/{viewerId}": {
			"get": {
				"operationId": "GetFollowings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/follows/counts/{userId}": {
			"get": {
				"operationId": "GetFollowCounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"followings": {},
										"followers": {}
									},
									"required": [
										"followings",
										"followers"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/follows/status/{userId}": {
			"get": {
				"operationId": "IsFollowing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isFollowing": {}
									},
									"required": [
										"isFollowing"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/auth/signup": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignUpDto"
							}
						}
					}
				}
			}
		},
		"/auth/signup/check-id": {
			"get": {
				"operationId": "CheckId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CheckIdResponse"
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/signin": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignInResponseDto"
								},
								"examples": {
									"Example 1": {
										"value": {
											"username": "주라미",
											"password": "00000000"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignInDto"
							}
						}
					}
				}
			}
		},
...
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}